Restructure the next catalog types and files a bit
