+++ .changeset/empty-berries-melt.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-catalog-backend': patch
+---
+
+Restructure the next catalog types and files a bit
