+++ plugins/catalog-backend/src/next/DefaultCatalogProcessingEngine.test.ts
@@ -15,12 +15,11 @@
  */
 
 import { getVoidLogger } from '@backstage/backend-common';
-import { DefaultProcessingDatabase } from './database/DefaultProcessingDatabase';
-
-import { DefaultCatalogProcessingEngine } from './DefaultCatalogProcessingEngine';
-import { Stitcher } from './Stitcher';
-import { CatalogProcessingOrchestrator } from './types';
 import waitForExpect from 'wait-for-expect';
+import { DefaultProcessingDatabase } from './database/DefaultProcessingDatabase';
+import { DefaultCatalogProcessingEngine } from './DefaultCatalogProcessingEngine';
+import { CatalogProcessingOrchestrator } from './processing/types';
+import { Stitcher } from './stitching/Stitcher';
 
 describe('DefaultCatalogProcessingEngine', () => {
   const db = ({
