+++ plugins/catalog-backend/src/next/processing/index.ts
@@ -14,7 +14,9 @@
  * limitations under the License.
  */
 
-export { NextCatalogBuilder } from './NextCatalogBuilder';
-export { createNextRouter } from './NextRouter';
-export * from './processing';
-export * from './stitching';
+export type {
+  CatalogProcessingOrchestrator,
+  EntityProcessingRequest,
+  EntityProcessingResult,
+} from './types';
+export { DefaultCatalogProcessingOrchestrator } from './DefaultCatalogProcessingOrchestrator';
