+++ plugins/catalog-backend/src/next/processing/types.ts
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2021 Spotify AB
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { Entity, EntityRelationSpec } from '@backstage/catalog-model';
+import { JsonObject } from '@backstage/config';
+
+export type EntityProcessingRequest = {
+  entity: Entity;
+  state: Map<string, JsonObject>; // Versions for multiple deployments etc
+};
+
+export type EntityProcessingResult =
+  | {
+      ok: true;
+      state: Map<string, JsonObject>;
+      completedEntity: Entity;
+      deferredEntities: Entity[];
+      relations: EntityRelationSpec[];
+      errors: Error[];
+    }
+  | {
+      ok: false;
+      errors: Error[];
+    };
+
+export interface CatalogProcessingOrchestrator {
+  process(request: EntityProcessingRequest): Promise<EntityProcessingResult>;
+}
