+++ plugins/catalog-backend/src/next/stitching/index.ts
@@ -14,7 +14,4 @@
  * limitations under the License.
  */
 
-export { NextCatalogBuilder } from './NextCatalogBuilder';
-export { createNextRouter } from './NextRouter';
-export * from './processing';
-export * from './stitching';
+export {};
